+++ taf/repository.py
@@ -1,4 +1,0 @@
-class BaseRepository:
-    def __init__(self, repo_path, repo_name):
-        self.repo_path = repo_path
-        self.repo_name = repo_name